refactor proofing code to proofing.py for signup
